Struct password_hash::ParamsString

pub struct ParamsString(_);

Algorithm parameter string.

The PHC string format specification defines a set of optional algorithm-specific name/value pairs which can be encoded into a PHC-formatted parameter string as follows:

$<param>=<value>(,<param>=<value>)*

This type represents that set of parameters.

Implementations

impl ParamsString

pub fn new() -> Self

Create new empty ParamsString.

pub fn add_b64_bytes<'a>( &mut self, name: impl TryInto<Ident<'a>>, bytes: &[u8] ) -> Result<()>

Add the given byte value to the ParamsString, encoding it as “B64”.

pub fn add_decimal<'a>( &mut self, name: impl TryInto<Ident<'a>>, value: Decimal ) -> Result<()>

Add a key/value pair with a decimal value to the ParamsString.

pub fn add_str<'a>( &mut self, name: impl TryInto<Ident<'a>>, value: impl TryInto<Value<'a>> ) -> Result<()>

Add a key/value pair with a string value to the ParamsString.

pub fn as_bytes(&self) -> &[u8]

Borrow the contents of this ParamsString as a byte slice.

pub fn as_str(&self) -> &str

Borrow the contents of this ParamsString as a str.

pub fn len(&self) -> usize

Get the count of the number ASCII characters in this ParamsString.

pub fn is_empty(&self) -> bool

Is this set of parameters empty?

pub fn iter(&self) -> Iter<'_>

Iterate over the parameters.

pub fn get<'a>(&self, name: impl TryInto<Ident<'a>>) -> Option<Value<'_>>

Get a parameter Value by name.

pub fn get_str<'a>(&self, name: impl TryInto<Ident<'a>>) -> Option<&str>

Get a parameter as a str.

pub fn get_decimal<'a>(&self, name: impl TryInto<Ident<'a>>) -> Option<Decimal>

Get a parameter as a Decimal.

See Value::decimal for format information.