password_hash

Module value

Algorithm parameter value as defined by the [PHC string format].

Implements the following parts of the specification:

The value for each parameter consists in characters in: [a-zA-Z0-9/+.-] (lowercase letters, uppercase letters, digits, /, +, . and -). No other character is allowed. Interpretation of the value depends on the parameter and the function. The function specification MUST unambiguously define the set of valid parameter values. The function specification MUST define a maximum length (in characters) for each parameter. For numerical parameters, functions SHOULD use plain decimal encoding (other encodings are possible as long as they are clearly defined).

Structs

Value

Algorithm parameter value string.

Functions

assert_valid_value 🔒

Are all of the given bytes allowed in a Value?

is_char_valid 🔒

Ensure the given ASCII character (i.e. byte) is allowed in a Value.

Type Aliases

Decimal

Type used to represent decimal (i.e. integer) values.